may 6

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jel-ghna <jel-ghna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 21:54:07 by josefelghna       #+#    #+#             */
-/*   Updated: 2025/04/22 12:56:12 by jel-ghna         ###   ########.fr       */
+/*   Updated: 2025/04/30 14:31:17 by jel-ghna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
